Time: 9 ms (59.03%), Space: 10.9 MB (64.39%) - LeetHub

+++ 0080-remove-duplicates-from-sorted-array-ii/0080-remove-duplicates-from-sorted-array-ii.c
@@ -1,7 +1,7 @@
 int removeDuplicates(int* nums, int numsSize) {
     bool couple_on = false;
-    int counter = 0, k = 1;
-    for (int i = 1; i < numsSize; i++) 
+    short counter = 0, k = 1;
+    for (int i = 1; i < (short)numsSize; i++) 
     {
         if (nums[i] == nums[i - 1] && couple_on ) //dup num
         {
